+++ src/OpenTelemetry.Exporter.Console/ConsoleMetricExporter.cs
@@ -41,7 +41,7 @@ namespace OpenTelemetry.Exporter
                     {
                         if (resourceAttribute.Key.Equals("service.name"))
                         {
-                            this.WriteLine("Service.Name" + resourceAttribute.Value);
+                            this.WriteLine("Service.Name=" + resourceAttribute.Value);
                         }
                     }
                 }
@@ -51,13 +51,13 @@ namespace OpenTelemetry.Exporter
             {
                 var msg = new StringBuilder($"\nExport ");
                 msg.Append(metric.Name);
-                if (!string.IsNullOrEmpty(metric.Description))
+                if (metric.Description != string.Empty)
                 {
                     msg.Append(", ");
                     msg.Append(metric.Description);
                 }
 
-                if (!string.IsNullOrEmpty(metric.Unit))
+                if (metric.Unit != string.Empty)
                 {
                     msg.Append($", Unit: {metric.Unit}");
                 }