+++ src/OpenTelemetry.Exporter.Prometheus/PrometheusExporter.cs
@@ -29,7 +29,6 @@ namespace OpenTelemetry.Exporter
     {
         internal const string HttpListenerStartFailureExceptionMessage = "PrometheusExporter http listener could not be started.";
         internal readonly PrometheusExporterOptions Options;
-        internal Batch<Metric> Metrics; // TODO: this is no longer needed, we can remove it later
         private readonly PrometheusExporterHttpServer metricsHttpServer;
         private Func<int, bool> funcCollect;
         private Func<Batch<Metric>, ExportResult> funcExport;