+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.0.0-rc9
+
+Released 2022-Feb-02
+
 * Fixed an issue with `Filter` and `Enrich` callbacks not firing under certain
   conditions when gRPC is used
   ([#2698](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2698))